requirements.txt
google_auth_oauthlib==0.4.1
google_api_python_client==1.12.2
pandas==1.1.2
protobuf==3.13.0
